+++ src/cmd/append.go
@@ -10,8 +10,8 @@ import (
 	"github.com/git-town/git-town/v9/src/step"
 	"github.com/git-town/git-town/v9/src/steps"
 	"github.com/git-town/git-town/v9/src/validate"
-	"github.com/git-town/git-town/v9/src/vm/interpreter"
-	"github.com/git-town/git-town/v9/src/vm/runstate"
+	runvm "github.com/git-town/git-town/v9/src/vm/interpreter"
+	"github.com/git-town/git-town/v9/src/vm/state"
 	"github.com/spf13/cobra"
 )
 
@@ -58,7 +58,7 @@ func executeAppend(arg string, debug bool) error {
 	if err != nil || exit {
 		return err
 	}
-	runState := runstate.RunState{
+	runState := state.RunState{
 		Command:             "append",
 		InitialActiveBranch: initialBranchesSnapshot.Active,
 		RunSteps:            appendSteps(config),
