+++ src/cmd/prune_branches.go
@@ -12,8 +12,8 @@ import (
 	"github.com/git-town/git-town/v9/src/messages"
 	"github.com/git-town/git-town/v9/src/step"
 	"github.com/git-town/git-town/v9/src/steps"
-	"github.com/git-town/git-town/v9/src/vm/interpreter"
-	"github.com/git-town/git-town/v9/src/vm/runstate"
+	runvm "github.com/git-town/git-town/v9/src/vm/interpreter"
+	"github.com/git-town/git-town/v9/src/vm/state"
 	"github.com/spf13/cobra"
 )
 
@@ -53,7 +53,7 @@ func executePruneBranches(debug bool) error {
 	if err != nil || exit {
 		return err
 	}
-	runState := runstate.RunState{
+	runState := state.RunState{
 		Command:             "prune-branches",
 		InitialActiveBranch: initialBranchesSnapshot.Active,
 		RunSteps:            pruneBranchesSteps(config, repo.Runner.Backend),
