+++ src/vm/persistence/save.go
@@ -8,11 +8,11 @@ import (
 
 	"github.com/git-town/git-town/v9/src/domain"
 	"github.com/git-town/git-town/v9/src/messages"
-	"github.com/git-town/git-town/v9/src/vm/runstate"
+	"github.com/git-town/git-town/v9/src/vm/state"
 )
 
 // Save stores the given run state for the given Git repo to disk.
-func Save(runState *runstate.RunState, repoDir domain.RepoRootDir) error {
+func Save(runState *state.RunState, repoDir domain.RepoRootDir) error {
 	content, err := json.MarshalIndent(runState, "", "  ")
 	if err != nil {
 		return fmt.Errorf(messages.RunstateSerializeProblem, err)
