+++ src/vm/state/core.go
@@ -1,4 +1,4 @@
-// Package runstate represents the current execution status of a Git Town command.
+// package state represents the current execution status of a Git Town command.
 // Runstates are created by the cmd package, executed by the runvm package,
 // and persisted by the persistence package.
-package runstate
+package state
